+++ src/Models/Transactions/CreateTransactionRequest.cs
@@ -9,9 +9,9 @@ namespace epay3.Sdk.Models
     {
         /// <summary>
         /// Gets or sets the total amount to charge not including any payer fees.
-        /// Required.
+/// Required.
         /// </summary>
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the payer shown on the receipt.
@@ -28,8 +28,8 @@ namespace epay3.Sdk.Models
         /// <summary>
         /// Gets or sets the pre-calculated payer fee.
         /// If set, the fee will not be re-calculated.
-        /// </summary>
-        public double? PayerFee { get; set; }
+ /// </summary>
+     public decimal? PayerFee { get; set; }
 
         /// <summary>
         /// Gets or sets comments shown on the receipt.
@@ -64,8 +64,8 @@ namespace epay3.Sdk.Models
         /// <summary>
         /// Gets or sets the fee charged by the initiating party.
         /// This does not include standard transaction fees.
-        /// </summary>
-        public double? InitiatingPartyFee { get; set; }
+     /// </summary>
+  public decimal? InitiatingPartyFee { get; set; }
 
         /// <summary>
         /// Gets or sets the IP address of the payer.
